# shared/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntakeResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IntakeData"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "IntakeResponse",
        "type": "object"
      },
...
      "TriageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TriageData"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "TriageResponse",
        "type": "object"
      },
...
      "VacancyDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VacancyItem"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "VacancyDetailResponse",
        "type": "object"
      },
      "VacancyHydrationResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VacancyItem"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "VacancyHydrationResponse",
        "type": "object"
      },
      "VacancyIntakeV1": {
        "additionalProperties": false,
        "properties": {
          "capture_source": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Source"
          },
          "captured_at": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "city": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "company_id": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "company_name": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "currency": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 12000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "experience": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience"
          },
          "parser_version": {
            "anyOf": [
              {
                "maxLength": 32,
                "pattern": "^[0-9]+(?:[.-][0-9A-Za-z-]+)*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Skills",
            "type": "array"
          },
          "source": {
            "maxLength": 32,
            "pattern": "^[a-zA-Z][a-zA-Z0-9_-]{0,31}$",
            "title": "Source",
            "type": "string"
          },
          "source_vacancy_id": {
            "default": "",
            "maxLength": 128,
            "title": "Source Vacancy Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "work_mode": {
            "anyOf": [
              {
                "enum": [
                  "remote",
                  "hybrid",
                  "office",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Mode"
          }
        },
        "required": [
          "schema_version",
          "source"
        ],
        "title": "VacancyIntakeV1",
        "type": "object"
      },
...
      "VacancyListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/VacancyItem"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/VacancyListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "VacancyListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-VacancyPilot-Client",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "VacancyPilot Ops Companion",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/vacancies": {
      "get": {
        "description": "List vacancies with bounded server-side filters and stable pagination.",
        "operationId": "vacancy_list_api_v1_vacancies_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "pattern": "^[a-zA-Z][a-zA-Z0-9_-]{0,31}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "work_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "remote",
                    "hybrid",
                    "office",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Mode"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Archived"
            }
          },
          {
            "in": "query",
            "name": "updated_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated After"
            }
          },
          {
            "in": "query",
            "name": "search_profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Vacancy List",
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/v1/vacancies/intake": {
      "post": {
        "description": "Idempotently capture a normalized vacancy into the companion.",
        "operationId": "vacancy_intake_api_v1_vacancies_intake_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-VacancyPilot-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Vacancypilot-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyIntakeV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Vacancy Intake",
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}": {
      "get": {
        "description": "Return a single vacancy by internal id.",
        "operationId": "vacancy_detail_api_v1_vacancies__vacancy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "title": "Vacancy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Vacancy Detail",
        "tags": [
          "vacancies"
        ]
      }
    },
...
    "/api/v1/vacancies/{vacancy_id}/hydrate": {
      "post": {
        "description": "Explicitly refresh one HH vacancy using GET /vacancies/{source_id}.",
        "operationId": "vacancy_hydrate_api_v1_vacancies__vacancy_id__hydrate_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "title": "Vacancy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyHydrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Vacancy Hydrate",
        "tags": [
          "vacancies"
        ]
      }
    },
    "/api/v1/vacancies/{vacancy_id}/triage": {
      "post": {
        "description": "Run deterministic no-LLM Stage A triage for a stored vacancy.",
        "operationId": "vacancy_triage_api_v1_vacancies__vacancy_id__triage_post",
        "parameters": [
          {
            "in": "path",
            "name": "vacancy_id",
            "required": true,
            "schema": {
              "title": "Vacancy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Vacancy Triage",
        "tags": [
          "vacancies"
        ]
      }
    }
  }
}